Support new location for apt repo
Node this is a cherry pick of certain parts of
https://github.com/puppetlabs/puppetlabs-kubernetes/pull/657

This is stalled due to yum issues and so doing this quick fix for
now to get us going. Once above is merged we can switch back
to upstream.

(cherry picked from commit 07acbc13d6ce622daaacb9af02250b1eda84c431)
Change-Id: I452a7e087cda5bb0089b5faa7b14c6c633ed1497